WIP for team aware middleware
